fix: use cursor to differentiate between move and resize events
Using event hooks, it is impossible to distinguish between move and
resize events directly. The previous method checked the change in
position and size to understand the type of event, but this is
problematic when windows force some minimum dimensions during
move/resize operations.

Now the style of the cursor is checked at the beginning of the operation
to understand whether the user's intent is to move or to resize the
window.