fix: prevent Windows Start menu from opening when setting window focus
For SetFocus to work, an input event must first be sent with SendInput.
Previously, the default input event sent before SetFocus activated the Windows Start menu
if the Win key was held down.

By using VK_NONAME as the virtual key, this problem no longer occurs.
This also solves the inability to use the Win key as the only modifier in key bindings.