fix(tm): use visible area as the tiling manager workspace
Previously, the monitor's workarea (which didn't account for custom
border padding) was used as the base of the tiling manager.
While this worked in most cases, it caused potential issues with large
paddings combined with many windows, as the window coordinates visible
to the user were different from those used by the tiling manager.